# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hika.ai API specification - AI Meeting Assistant
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Session {
  id: number;
  title: string;
  /** zoom | teams | meet | other */
  platform: string;
  /** active | ended */
  status?: string;
  insightCount?: number;
  createdAt: string;
  /** @nullable */
  endedAt?: string | null;
}

export interface SessionInput {
  /** @minLength 1 */
  title: string;
  platform: string;
}

export interface SessionUpdate {
  title?: string;
  status?: string;
}

export interface Insight {
  id: number;
  sessionId: number;
  question: string;
  answer: string;
  /** @nullable */
  confidence?: string | null;
  createdAt: string;
}

export interface InsightInput {
  sessionId: number;
  question: string;
  answer: string;
  confidence?: string;
}

export interface Stats {
  totalSessions: number;
  activeSessions: number;
  totalInsights: number;
  avgInsightsPerSession: number;
}

export type AnalyzeInputMode = typeof AnalyzeInputMode[keyof typeof AnalyzeInputMode];


export const AnalyzeInputMode = {
  interview: 'interview',
  meeting: 'meeting',
} as const;

export type AnalyzeInputUploadedDocsItem = {
  id: string;
  name: string;
};

export type ConversationTurnRole = typeof ConversationTurnRole[keyof typeof ConversationTurnRole];


export const ConversationTurnRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ConversationTurn {
  role: ConversationTurnRole;
  /** @maxLength 4000 */
  content: string;
}

export interface AnalyzeInput {
  /**
     * Recent meeting transcript text
     * @maxLength 12000
     */
  transcript: string;
  /**
     * Optional base64-encoded PNG screenshot of the meeting screen
     * @nullable
     */
  screenshotBase64?: string | null;
  /** @nullable */
  sessionId?: number | null;
  mode?: AnalyzeInputMode;
  /** Requested model; the API enforces its server-side allowlist */
  model?: string;
  /** @maxItems 12 */
  history?: ConversationTurn[];
  /**
     * Optional uploaded documents to use as grounded context
     * @maxItems 3
     */
  uploadedDocs?: AnalyzeInputUploadedDocsItem[];
}

export interface AISection {
  /** technical|rootCause|solution|code|sql|pyspark|azure|architecture|bestPractices|risks|businessImpact */
  type: string;
  title: string;
  content: string;
  /** @nullable */
  language?: string | null;
}

export type AnalyzeResultConfidence = typeof AnalyzeResultConfidence[keyof typeof AnalyzeResultConfidence];


export const AnalyzeResultConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface AnalyzeResult {
  /** The question or topic detected in the transcript that was answered */
  question: string;
  answer: string;
  questionType?: string;
  /** Detected response domain */
  domain?: string;
  suggestions: string[];
  confidence: AnalyzeResultConfidence;
  sections?: AISection[];
}

export interface TranscribeInput {
  /** Base64-encoded audio (webm/ogg format) */
  audioBase64: string;
  mimeType?: string;
}

export interface TranscribeResult {
  transcript: string;
}