# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hika.ai API specification - AI Meeting Assistant
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /insights:
    post:
      operationId: createInsight
      tags: [insights]
      summary: Store an AI-generated insight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsightInput"
      responses:
        "201":
          description: Created insight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insight"
...
components:
  schemas:
...
    Insight:
      type: object
      required: [id, sessionId, question, answer, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        question:
          type: string
        answer:
          type: string
        confidence:
          type: ["string", "null"]
        createdAt:
          type: string
    InsightInput:
      type: object
      required: [sessionId, question, answer]
      properties:
        sessionId:
          type: integer
        question:
          type: string
        answer:
          type: string
        confidence:
          type: string
...